requirements.txt
mcp[cli]
requests
biopython
python-dotenv
python-json-logger